Promote IPv6 dhcpcd address flags to unsignned int to allow expansion.

+++ src/ipv6nd.h
@@ -90,9 +90,9 @@ void ipv6nd_printoptions(const struct dhcpcd_ctx *,
 void ipv6nd_startrs(struct interface *);
 ssize_t ipv6nd_env(char **, const char *, const struct interface *);
 const struct ipv6_addr *ipv6nd_iffindaddr(const struct interface *ifp,
-    const struct in6_addr *addr, short flags);
+    const struct in6_addr *addr, unsigned int flags);
 struct ipv6_addr *ipv6nd_findaddr(struct dhcpcd_ctx *,
-    const struct in6_addr *, short);
+    const struct in6_addr *, unsigned int);
 void ipv6nd_freedrop_ra(struct ra *, int);
 #define ipv6nd_free_ra(ra) ipv6nd_freedrop_ra((ra),  0)
 #define ipv6nd_drop_ra(ra) ipv6nd_freedrop_ra((ra),  1)
